+++ src/test/resources/addOperation.rule
@@ -1,0 +1,17 @@
+{
+    "type": "and",
+    "children": [
+        {
+            "type": "equals",
+            "path": "/value",
+            "value": 11,
+            "preoperation": {"operation" : "add", "operand" : 3}
+        },
+        {
+            "type": "equals",
+            "path": "/value",
+            "value": 30,
+            "preoperation": {"operation" : "add", "operand" : 22}
+        }
+    ]
+}
