+++ src/test/resources/divideOperation.rule
@@ -1,0 +1,17 @@
+{
+    "type": "and",
+    "children": [
+        {
+            "type": "equals",
+            "path": "/value",
+            "value": 4,
+            "preoperation": {"operation" : "divide", "operand" : 2}
+        },
+        {
+            "type": "less_than",
+            "path": "/value",
+            "value": 5,
+            "preoperation": {"operation" : "divide", "operand" : 2}
+        }
+    ]
+}
