+++ src/test/resources/epochOperation.rule
@@ -1,0 +1,18 @@
+{
+    "type": "and",
+    "children": [
+        {
+            "type": "greater_than_equals",
+            "path": "/unixTime",
+            "value": 11,
+            "preoperation": {"operation" : "epoch", "operand" : "hour_of_day", "zoneOffSet":"+05:30"}
+            
+        },
+        {
+            "type": "less_than_equals",
+            "path": "/unixTime",
+            "value": 3,
+            "preoperation": {"operation" : "epoch", "operand" : "day_of_week", "zoneOffSet":"+05:30"}
+        }
+    ]
+}
