+++ src/test/resources/multiplyOperation.rule
@@ -1,0 +1,17 @@
+{
+    "type": "and",
+    "children": [
+        {
+            "type": "greater_than",
+            "path": "/value",
+            "value": 11,
+            "preoperation": {"operation" : "multiply", "operand" : 2}
+        },
+        {
+            "type": "less_than",
+            "path": "/value",
+            "value": 30,
+            "preoperation": {"operation" : "multiply", "operand" : 3}
+        }
+    ]
+}
